requirements.txt
python-dateutil
asyncio
aiohttp
aiokafka
requests
kafka-python
docker
boto3
PyYAML
celery
redis
httpx
redis-om
fastapi
#jwt
passlib
uvicorn
bcrypt
psutil
grpcio
grpcio-tools
protobuf
#celery-beat